Rename AWS authentication mechanism to AWS-IAM
