+++ src/protocol/sasl/awsIam/index.js
