+++ src/protocol/sasl/awsIam/request.js
