+++ src/protocol/sasl/awsIam/response.js
